requirements.txt
qrcode>=7.4.2,<7.5
pillow>=10.4.0,<10.5